+++ .idea/workspace.xml
@@ -1,0 +1,119 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project version="4">
+  <component name="AutoImportSettings">
+    <option name="autoReloadType" value="SELECTIVE" />
+  </component>
+  <component name="ChangeListManager">
+    <list default="true" id="0a6294ed-ab28-4f72-aa8d-e2f022986984" name="Changes" comment="add: 테스트를 위한 test 폴더 및 하위 테스트 파일 추가, 메인 Calculator.java에서 연산 로직만 분리해서 CalculatorService.java로 생성">
+      <change afterPath="$PROJECT_DIR$/.idea/misc.xml" afterDir="false" />
+      <change afterPath="$PROJECT_DIR$/.idea/workspace.xml" afterDir="false" />
+      <change afterPath="$PROJECT_DIR$/out/production/Calculator/main/Calculator$4.class" afterDir="false" />
+      <change afterPath="$PROJECT_DIR$/out/production/Calculator/main/CalculatorService.class" afterDir="false" />
+      <change afterPath="$PROJECT_DIR$/out/test/Calculator/main/CalculatorServiceTest.class" afterDir="false" />
+      <change afterPath="$PROJECT_DIR$/out/test/Calculator/main/JButtonMemoryTest.class" afterDir="false" />
+      <change afterPath="$PROJECT_DIR$/out/test/Calculator/main/JButtonSTest.class" afterDir="false" />
+      <change afterPath="$PROJECT_DIR$/out/test/Calculator/main/JButtonWhiteTest.class" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/Calculator.java" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/CalculatorBtnActionListener.java" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/Hello.java" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/README.md" beforeDir="false" afterPath="$PROJECT_DIR$/README.md" afterDir="false" />
+    </list>
+    <option name="SHOW_DIALOG" value="false" />
+    <option name="HIGHLIGHT_CONFLICTS" value="true" />
+    <option name="HIGHLIGHT_NON_ACTIVE_CHANGELIST" value="false" />
+    <option name="LAST_RESOLUTION" value="IGNORE" />
+  </component>
+  <component name="FileTemplateManagerImpl">
+    <option name="RECENT_TEMPLATES">
+      <list>
+        <option value="Class" />
+        <option value="JUnit5 Test Class" />
+      </list>
+    </option>
+  </component>
+  <component name="Git.Settings">
+    <option name="RECENT_GIT_ROOT_PATH" value="$PROJECT_DIR$" />
+  </component>
+  <component name="MarkdownSettingsMigration">
+    <option name="stateVersion" value="1" />
+  </component>
+  <component name="ProjectId" id="2tGD6PyrQTiT9ZrgUT0tZjEB5W7" />
+  <component name="ProjectLevelVcsManager" settingsEditedManually="true" />
+  <component name="ProjectViewState">
+    <option name="hideEmptyMiddlePackages" value="true" />
+    <option name="showLibraryContents" value="true" />
+  </component>
+  <component name="PropertiesComponent"><![CDATA[{
+  "keyToString": {
+    "Downloaded.Files.Path.Enabled": "false",
+    "Repository.Attach.Annotations": "false",
+    "Repository.Attach.JavaDocs": "false",
+    "Repository.Attach.Sources": "false",
+    "RunOnceActivity.OpenProjectViewOnStart": "true",
+    "RunOnceActivity.ShowReadmeOnStart": "true",
+    "SHARE_PROJECT_CONFIGURATION_FILES": "true",
+    "com.intellij.testIntegration.createTest.CreateTestDialog.defaultLibrary": "JUnit5",
+    "com.intellij.testIntegration.createTest.CreateTestDialog.defaultLibrarySuperClass.JUnit5": "",
+    "project.structure.last.edited": "Modules",
+    "project.structure.proportion": "0.15",
+    "project.structure.side.proportion": "0.2",
+    "settings.editor.selected.configurable": "reference.settings.ide.settings.file-colors"
+  }
+}]]></component>
+  <component name="RecentsManager">
+    <key name="CreateTestDialog.Recents.Supers">
+      <recent name="" />
+    </key>
+    <key name="CreateTestDialog.RecentsKey">
+      <recent name="main" />
+    </key>
+  </component>
+  <component name="RunManager">
+    <configuration default="true" type="JetRunConfigurationType">
+      <method v="2">
+        <option name="Make" enabled="true" />
+      </method>
+    </configuration>
+    <configuration default="true" type="KotlinStandaloneScriptRunConfigurationType">
+      <option name="filePath" />
+      <method v="2">
+        <option name="Make" enabled="true" />
+      </method>
+    </configuration>
+  </component>
+  <component name="SpellCheckerSettings" RuntimeDictionaries="0" Folders="0" CustomDictionaries="0" DefaultDictionary="application-level" UseSingleDictionary="true" transferred="true" />
+  <component name="TaskManager">
+    <task active="true" id="Default" summary="Default task">
+      <changelist id="0a6294ed-ab28-4f72-aa8d-e2f022986984" name="Changes" comment="" />
+      <created>1739969342949</created>
+      <option name="number" value="Default" />
+      <option name="presentableId" value="Default" />
+      <updated>1739969342949</updated>
+    </task>
+    <task id="LOCAL-00001" summary="add: 테스트를 위한 test 폴더 및 하위 테스트 파일 추가, 메인 Calculator.java에서 연산 로직만 분리해서 CalculatorService.java로 생성">
+      <option name="closed" value="true" />
+      <created>1739974126047</created>
+      <option name="number" value="00001" />
+      <option name="presentableId" value="LOCAL-00001" />
+      <option name="project" value="LOCAL" />
+      <updated>1739974126047</updated>
+    </task>
+    <option name="localTasksCounter" value="2" />
+    <servers />
+  </component>
+  <component name="Vcs.Log.Tabs.Properties">
+    <option name="TAB_STATES">
+      <map>
+        <entry key="MAIN">
+          <value>
+            <State />
+          </value>
+        </entry>
+      </map>
+    </option>
+  </component>
+  <component name="VcsManagerConfiguration">
+    <MESSAGE value="add: 테스트를 위한 test 폴더 및 하위 테스트 파일 추가, 메인 Calculator.java에서 연산 로직만 분리해서 CalculatorService.java로 생성" />
+    <option name="LAST_COMMIT_MESSAGE" value="add: 테스트를 위한 test 폴더 및 하위 테스트 파일 추가, 메인 Calculator.java에서 연산 로직만 분리해서 CalculatorService.java로 생성" />
+  </component>
+</project>
