--- Hello.java
@@ -1,20 +1,0 @@
-System.out.println("Hello, World!");
-System.out.println("Hello, World!");
-System.out.println("Hello, World!");
-System.out.println("Hello, World!");
-System.out.println("Hello, World!");
-System.out.println("Hello, World!");
-System.out.println("Hello, World!");
-System.out.println("Hello, World!");
-System.out.println("Hello, World!");
-System.out.println("Hello, World!");
-System.out.println("Hello, World!");
-System.out.println("Hello, World!");
-System.out.println("Hello, World!");
-System.out.println("Hello, World!");
-System.out.println("Hello, World!");
-System.out.println("Hello, World!");
-System.out.println("Hello, World!");
-System.out.println("Hello, World!");
-System.out.println("Hello, World!");
-System.out.println("Hello, World!");
