+++ README.md
@@ -35,7 +35,7 @@
 
 2. Java 개발 환경을 설정합니다. (Java 8 이상 권장)
 
-3. IDE에서 `Calculator` 클래스를 실행하여 애플리케이션을 시작합니다.
+3. IDE에서 `java.Calculator` 클래스를 실행하여 애플리케이션을 시작합니다.
 
 ## 기여 방법
