+++ src/main/Main.java
@@ -1,0 +1,7 @@
+package main;
+
+public class Main {
+    public static void main(String[] args) {
+        new Calculator();
+    }
+}
